moved find manifest functions into forc util
