+++ forc/src/ops/forc_clean.rs
@@ -1,8 +1,8 @@
 use crate::cli::CleanCommand;
 use anyhow::{anyhow, bail, Result};
-use forc_util::default_output_directory;
+use forc_util::{default_output_directory, find_cargo_manifest_dir, find_manifest_dir};
 use std::{path::PathBuf, process};
-use sway_utils::{find_cargo_manifest_dir, find_manifest_dir, MANIFEST_FILE_NAME};
+use sway_utils::MANIFEST_FILE_NAME;
 
 pub fn clean(command: CleanCommand) -> Result<()> {
     let CleanCommand { path } = command;
