+++ sway-lsp/Cargo.toml
@@ -10,6 +10,7 @@ description = "LSP server for Sway."
 
 [dependencies]
 dashmap = "4.0.2"
+forc-util = { version = "0.6.1", path = "../forc-util" }
 tower-lsp = "0.16.0"
 ropey = "1.2"
 serde_json = "1.0.60"
